+++ src/components/team/team-item.js
@@ -11,34 +11,49 @@ import VerifiedBadge from 'Components/verified-badge';
 import ProfileList from 'Components/profile-list';
 import { WrappingLink } from 'Components/link';
 import { getLink, getAvatarUrl, DEFAULT_TEAM_AVATAR } from 'Models/team';
+import { createAPIHook } from 'State/api';
+import { captureException } from 'Utils/sentry';
 
 import styles from './team-item.styl';
+
+const useTeamUsers = createAPIHook(async (api, teamID) => {
+  try {
+    const res = await api.get(`/v1/teams/by/id/users?id=${teamID}`);
+    return res.data.items;
+  } catch (e) {
+    captureException(e);
+    return [];
+  }
+});
 
 const ProfileAvatar = ({ team }) => <Image className={styles.avatar} src={getAvatarUrl(team)} defaultSrc={DEFAULT_TEAM_AVATAR} alt="" />;
 
 const getTeamThanksCount = (team) => sumBy(team.users, (user) => user.thanksCount);
 
-const TeamItem = ({ team }) => (
-  <WrappingLink className={styles.container} href={getLink(team)}>
-    <Cover type="team" item={team} size="medium" />
-    <div className={styles.mainContent}>
-      <div className={styles.avatarWrap}>
-        <ProfileAvatar team={team} />
-      </div>
-      <div className={styles.body}>
-        <div className={styles.itemButtonWrap}>
-          <Button href={getLink(team)}>{team.name}</Button>
-          {!!team.isVerified && <VerifiedBadge />}
+const TeamItem = ({ team }) => {
+  const { value: users } = useTeamUsers(team.id);
+  return (
+    <WrappingLink className={styles.container} href={getLink(team)}>
+      <Cover type="team" item={team} size="medium" />
+      <div className={styles.mainContent}>
+        <div className={styles.avatarWrap}>
+          <ProfileAvatar team={team} />
         </div>
-        <div className={styles.usersList}>
-          <ProfileList layout="block" users={team.users} />
+        <div className={styles.body}>
+          <div className={styles.itemButtonWrap}>
+            <Button href={getLink(team)}>{team.name}</Button>
+            {!!team.isVerified && <VerifiedBadge />}
+          </div>
+          <div className={styles.usersList}>
+            <ProfileList layout="block" users={users} />
+          </div>
+          <Markdown length={96}>{team.description || ' '}</Markdown>
+          <Thanks count={getTeamThanksCount(team)} />
         </div>
-        <Markdown length={96}>{team.description || ' '}</Markdown>
-        <Thanks count={getTeamThanksCount(team)} />
       </div>
-    </div>
-  </WrappingLink>
-);
+    </WrappingLink>
+  );
+};
 
 TeamItem.propTypes = {
   team: PropTypes.shape({
